Changed models, added coach owed endpoint, added to post event endpoint, updated seed.go and updated README (#27)
Updated the coach and event models to included the amount owed to a coach and the event expenses. Added an endpoint to get the amount owed to a coach. Added a GetOwed function and updated the PostEvent function to update the corresponding coach's amount owed on event creation. Updated the seed file to include the added fields in the coach and event models. Finally also updated the README since it was behind.